Extract dummy data and consolidate shared render utilities
- Move generate_dummy_data() from main.py to src/dummy_data.py, reducing
  main.py from 632 to ~420 lines of orchestration logic
- Consolidate duplicated utilities into src/render/primitives.py:
  - fmt_time(): was duplicated in 4 component files
  - wrap_lines(): was duplicated in qotd_panel and info_panel
  - events_for_day(): was duplicated in week_view and today_view
  - deg_to_compass(): moved from src/fetchers/weather.py to remove
    render→fetcher coupling (3 component files imported it)
